#77 feat: 녹화 종료 후 전달받은 studyRecordId로 분석 결과 조회

+++ feature/camera/src/main/java/com/ds/studify/feature/camera/navigation/CameraNavigation.kt
@@ -19,7 +19,7 @@ data object RouteCamera
 
 fun NavGraphBuilder.cameraScreen(
     navHostController: NavHostController,
-    navigateToAnalysis: () -> Unit
+    navigateToAnalysis: (Long) -> Unit
 ) {
     navigation<NavRouteCamera>(
         startDestination = RouteCameraGuide
